% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{gimmedata}
\alias{gimmedata}
\title{TSD Data reader}
\usage{
gimmedata(
  DataDir = getwd(),
  ExperimentID = "[0-9]{5}",
  ExperimentName = ".*",
  UniqueName = ".*",
  Session = ".*",
  Run = ".*",
  clean = T,
  verbose = F
)
}
\arguments{
\item{DataDir}{the root directory where all data_exp_* folders are found.}

\item{ExperimentID}{Experiment ID number to extract, as defined \href{https://docs.google.com/spreadsheets/d/1p6_WHQXNGFw2EJGny1jb5qivMy2pJ_VRRYoDGRLxgbY/edit#gid=0}{here}.}

\item{ExperimentName}{a country label, as defined \href{https://docs.google.com/spreadsheets/d/1p6_WHQXNGFw2EJGny1jb5qivMy2pJ_VRRYoDGRLxgbY/edit#gid=0}{here}.}

\item{UniqueName}{Task or Questionnaire name, as defined \href{https://docs.google.com/spreadsheets/d/1p6_WHQXNGFw2EJGny1jb5qivMy2pJ_VRRYoDGRLxgbY/edit#gid=0}{here}.}

\item{Session}{Session name (1, 2, 3, control...).}

\item{Run}{Run number (1, 2, 3).}

\item{clean}{whether or not to clean output from less usefull data (such as checkpoints, schedule ID etc.).}

\item{verbose}{whether or not to display debugging info.}
}
\value{
a tibble: data for a specific selection of data.
}
\description{

}
\details{
NOTE: the data can easily be passed to dplyr verbs.

This function looks for appropriately formatted data in directory DataDir.
}
\examples{
\dontrun{

   mydata <- gimmedata(ExperimentName = 'FR',UniqueName = 'Implicit')

   mydata <- gimmedata(ExperimentName = 'FR',UniqueName = 'Implicit') \%>\%
      filter(`Screen Name` == 'response')

}
}
